fixed build_bims_NAc_genes.sh
an earlier iteration of this script would actually generate wrapper scripts for different features and tissues/subregions of the brain. I made this script specific to NAc_gene, requested more memory and cores, and removed the stdout/stderr flags, because they were not working as intended for some reason (telling me that the logs could not be opened? when they should be getting written to file?) this should work