make bim snps unique
plink was giving me a warning that there were duplicate snps. I originally used unique() but that filters out elements in a vector, so Leo suggested I use make.names() to generate unique names for the snps. He also suggested that I do this earlier in the pipeline on the precursor input i.e. in filter_snps.R, but I couldn't figure out which object to work on. I might change this later if this turns out to be a problem.