keeping duplicate SNP IDs for filter_snps -> build_bims workflow
After struggling with build_bims.R, we've decided to go ahead and keep the duplicated snp IDs that were present in the plink output from filter_snps simply because build_bims.R does not work when we use make.names() at any point to generate unique snp IDs for the ouput bim file from filter_snps/input bim file for build_bims. Leo said he would revisit this on his own later.  I also updated the output path for filter_snps and the input for build_bims to reflect the two different sets we've generated (with and without duplicate RSIDs) so there is no confusion later with which bim file is which.